Refactor: Move Jetpack-related files to its own package.
