+++ WooCommerce/src/main/res/navigation/nav_graph_main.xml
@@ -258,7 +258,7 @@
         app:popExitAnim="@anim/activity_fade_out" />
     <dialog
         android:id="@+id/jetpackBenefitsDialog"
-        android:name="com.woocommerce.android.ui.mystore.JetpackBenefitsDialog"
+        android:name="com.woocommerce.android.ui.jetpack.JetpackBenefitsDialog"
         android:label="JetpackBenefitsDialog" >
         <action
             android:id="@+id/action_jetpackBenefitsDialog_to_jetpackInstallStartDialog"
@@ -266,7 +266,7 @@
     </dialog>
     <dialog
         android:id="@+id/jetpackInstallStartDialog"
-        android:name="com.woocommerce.android.ui.mystore.JetpackInstallStartDialog"
+        android:name="com.woocommerce.android.ui.jetpack.JetpackInstallStartDialog"
         android:label="JetpackInstallStartDialog"
         tools:layout="@layout/dialog_jetpack_install_start">
         <action
@@ -275,7 +275,7 @@
     </dialog>
     <dialog
         android:id="@+id/jetpackInstallProgressDialog"
-        android:name="com.woocommerce.android.ui.mystore.JetpackInstallProgressDialog"
+        android:name="com.woocommerce.android.ui.jetpack.JetpackInstallProgressDialog"
         android:label="JetpackInstallProgressDialog"
         tools:layout="@layout/dialog_jetpack_install_progress" >
         <action
